# api-gateway/openapi/public-api.yaml
openapi: 3.0.3
info:
  title: AutoVaultLK Public API
  description: |
    North-south routes per SAD section 3.5.1.
    JWT validation is performed by each NestJS service, not at the gateway.
  version: 1.0.0
servers:
  - url: https://{apiId}.execute-api.{region}.amazonaws.com/{stage}
    description: AWS API Gateway (production)
    variables:
      apiId:
        default: example
      region:
        default: ap-southeast-1
      stage:
        default: dev
  - url: http://localhost:8080
    description: Local nginx path proxy (development)
...
paths:
...
  /marketplace/listings:
    get:
      tags:
        - marketplace
      summary: Browse listings (public - LIVE only)
      operationId: listListings
      responses:
        '200':
          description: Paginated listings
    post:
      tags:
        - marketplace
      summary: Create listing (dealer)
      operationId: createListing
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /marketplace/listings/{id}:
    get:
      tags:
        - marketplace
      summary: Get listing by id (public if LIVE)
      operationId: getListing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Listing detail
    patch:
      tags:
        - marketplace
      summary: Update listing (dealer)
      operationId: updateListing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Updated
  /marketplace/listings/approve-all:
    patch:
      tags:
        - marketplace
      summary: Approve every PENDING_REVIEW listing the dealer owns (dealer)
      description: >
        Bulk form of the single-listing approve. Moves all of the calling
        dealer's PENDING_REVIEW listings to LIVE in one statement and returns
        how many moved. Scoped to the caller by the bearer token; `approved: 0`
        (not an error) when nothing was pending.
      operationId: approveAllPendingListings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Approved. `data.approved` is the number published.
        '401':
          description: No or invalid bearer token.
        '403':
          description: Not a DEALER.
  /marketplace/listings/{id}/approve:
    patch:
      tags:
        - marketplace
      summary: Approve a PENDING_REVIEW listing, publishing it (dealer)
      description: >
        FR-42/FR-42.1. Moves the listing from PENDING_REVIEW to LIVE. This is
        the explicit approval FR-42 requires before any ETL-loaded listing
        becomes publicly visible.
      operationId: approveListing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Approved and published.
        '401':
          description: No or invalid bearer token.
        '403':
          description: Not a DEALER/ADMIN, or a DEALER who does not own this listing.
        '404':
          description: No such listing.
        '409':
          description: The listing exists but is not PENDING_REVIEW - nothing to approve.
  /marketplace/listings/{id}/images:
    post:
      tags:
        - marketplace
      summary: Attach photos to a listing the caller owns (dealer)
      description: >
        FR-58. Replaces the listing's whole image set - a re-upload means
        "this is the current set of photos", not "add more to what's there".
        The first file becomes the primary photo (the one search/detail
        responses use for imageUrl/thumbnailUrl).


        Only available when the deployment's IMAGE_SERVE_MODE is s3 or
        local; in demo mode this 400s, since an upload that mode could never
        serve back would be worse than refusing it outright.
      operationId: uploadListingImages
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - images
              properties:
                images:
                  type: array
                  maxItems: 10
                  items:
                    type: string
                    format: binary
                  description: >
                    1-10 JPEG, PNG or WebP files, max 8 MB each. Order sets
                    display_order; the first file is stored as the primary
                    photo.
      responses:
        '201':
          description: Stored. Returns the vehicle_images rows created.
        '400':
          description: >
            No file attached, an unsupported file type, a file over 8 MB, more
            than 10 files, or the deployment is running in demo mode.
        '401':
          description: No or invalid bearer token.
        '403':
          description: Not a DEALER/ADMIN, or a DEALER who does not own this listing.
        '404':
          description: No such listing.
...
  /marketplace/search/filters:
    get:
      tags:
        - marketplace
      summary: Filter-based search over LIVE listings (public - guest browsing)
      operationId: filterSearch
      parameters:
        - name: vehicleType
          in: query
          schema:
            type: string
          description: Comma-separated, e.g. CAR,SUV
        - name: make
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: fuelType
          in: query
          schema:
            type: string
        - name: transmissionType
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: integer
        - name: minYear
          in: query
          schema:
            type: integer
        - name: maxYear
          in: query
          schema:
            type: integer
        - name: minMileage
          in: query
          schema:
            type: integer
        - name: maxMileage
          in: query
          schema:
            type: integer
        - name: verifiedDealersOnly
          in: query
          schema:
            type: boolean
        - name: hasRegistrationYear
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
          description: Keyword match against the tsvector search_vector column
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, year_desc, mileage_asc, newest]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: facets
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated, filtered listings
  /marketplace/search/nl:
    get:
      tags:
        - marketplace
      summary: Natural-language search over LIVE listings (public - guest browsing)
      description: |
        Deterministic parser (FR-21) turns `q` into structured filters. When
        parse confidence is below 0.6, unresolved tokens are sent to Groq
        and whitelist-validated (FR-21.2 / FR-21.3) before merging. Groq
        outages fall back to rules-only filters. Leftover semantic text is
        embedded with all-MiniLM-L6-v2 and ranked with
        `ORDER BY embedding <=> query_vector` (FR-22 / FR-23). If MiniLM is
        unavailable, leftovers are ranked with pg_trgm `word_similarity` on
        `search_text` (FR-24); a `search_text` WHERE is applied only when
        the parser extracted no filters. Resolved make/model are never
        re-matched via trigram.
      operationId: nlSearch
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            maxLength: 500
          description: Free-text query, e.g. "Toyata Corrola under 8.5m deisel"
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, year_desc, mileage_asc, newest]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: facets
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated listings plus parse metadata (confidence, unresolved tokens)
...
  /marketplace/search/options:
    get:
      tags:
        - marketplace
      summary: Dropdown option values - enums plus make/model dictionary (public)
      operationId: searchOptions
      parameters:
        - name: vehicleType
          in: query
          schema:
            type: string
          description: Scopes the make/model list to this vehicle type
      responses:
        '200':
          description: Vehicle types, conditions, fuel/transmission enums, body types, makes with nested models
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...